♻️ #181 refactor: notes 페이지 로딩 ui 변화 필요컴포넌트 최소화 작업
