+++ components/notes/NotesList.tsx
@@ -1,12 +1,13 @@
-import { Note } from '@/app/(nav)/notes/[goalId]/page';
+import { Note } from '@/lib/types/todo';
 import NoteItem from './NoteItem';
 
 interface NotesListProps {
   notes: Note[];
+  isFetching: boolean;
 }
 
-const NotesList: React.FC<NotesListProps> = ({ notes }) => {
-  if (!notes.length) {
+const NotesList: React.FC<NotesListProps> = ({ notes, isFetching }) => {
+  if (!notes.length && !isFetching) {
     return <div className='flex justify-center items-center text-sm text-slate-500'>아직 등록된 노트가 없어요</div>;
   }
   return (
